pyproject.toml
[tool.poetry]
name = "secret-tool"
version = "0.1.0"
description = "Secret tool demo"
authors = ["Adam Peacock <apeacock@atlantistech.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.15.0"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
rope = "^1.3.0"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
